+++ src/diff.ts
@@ -1,5 +1,5 @@
 
-import {GeoJSONSourceSpecification, LayerSpecification, LightSpecification, SkySpecification, SourceSpecification, SpriteSpecification, StyleSpecification, TerrainSpecification, TransitionSpecification} from './types.g';
+import {GeoJSONSourceSpecification, LayerSpecification, LightSpecification, ProjectionSpecification, SkySpecification, SourceSpecification, SpriteSpecification, StyleSpecification, TerrainSpecification, TransitionSpecification} from './types.g';
 import isEqual from './util/deep_equal';
 
 /**
@@ -28,6 +28,7 @@ export type DiffOperationsMap = {
     'setLight': [LightSpecification];
     'setTerrain': [TerrainSpecification];
     'setSky': [SkySpecification];
+    'setProjection': [ProjectionSpecification];
 }
 
 export type DiffOperations = keyof DiffOperationsMap;
@@ -303,6 +304,9 @@ function diffStyles(before: StyleSpecification, after: StyleSpecification): Diff
         }
         if (!isEqual(before.sky, after.sky)) {
             commands.push({command: 'setSky', args: [after.sky]});
+        }
+        if (!isEqual(before.projection, after.projection)) {
+            commands.push({command: 'setProjection', args: [after.projection]});
         }
 
         // Handle changes to `sources`